+++ src/test/kotlin/com/github/ekvedaras/classfactoryphpstorm/MakeMethodFactoryTest.kt
@@ -27,4 +27,16 @@ internal class MakeMethodFactoryTest : EssentialTestCase() {
     override fun testItResolvesReferencesInArrayReturnedByDirectlyPassedClosure() {
         // TODO: failing due to PhpCache having cached empty results. Find a workaround
     }
+
+    fun testItCanHandleMethodCallsOnOtherFactoriesFromAttributesArray() {
+        assertInspection(
+            "makeOtherFactoryFromAttributes.php",
+            incorrectPropertyTypeInspection(),
+        )
+
+        assertInspection(
+            "makeOtherFactoryFromAttributes.php",
+            incorrectPropertyTypeInClosureReturnsInspection(),
+        )
+    }
 }